+++ src/features/speedrun/season2/callbacks/postRender.ts
@@ -10,13 +10,10 @@ import {
   SEASON_2_LOCK_SECONDS,
   SEASON_2_NUM_BANS,
 } from "../constants";
-import sprites from "../sprites";
+import { drawSeason2StartingRoomSprites } from "../startingRoomSprites";
 import v from "../v";
 
 const TOP_LEFT_GRID_INDEX = 32;
-const TOP_RIGHT_GRID_INDEX = 42;
-const SPRITE_TITLE_OFFSET = Vector(0, -30);
-const SPRITE_ITEM_OFFSET = 15;
 
 export function season2PostRender(): void {
   const hud = game.GetHUD();
@@ -38,7 +35,7 @@ export function season2PostRender(): void {
     return;
   }
 
-  drawStartingRoomSprites();
+  drawSeason2StartingRoomSprites();
   drawStartingRoomText();
 }
 
@@ -69,13 +66,18 @@ function drawErrors() {
   return true;
 }
 
-function drawStartingRoomSprites() {
-  for (const [spriteName, sprite] of Object.entries(sprites)) {
-    if (sprite !== null) {
-      const position = getPosition(spriteName as keyof typeof sprites);
-      sprite.RenderLayer(0, position);
-    }
+function getSeason2ErrorMessage(action: string, secondsRemaining: int) {
+  if (secondsRemaining > SEASON_2_LOCK_SECONDS) {
+    return 'Please set your item vetos for Season 2 again in the "Change Char Order" custom challenge.';
   }
+
+  const suffix = secondsRemaining > 1 ? "s" : "";
+  const secondsRemainingText = `${secondsRemaining} second${suffix}`;
+  const secondSentence =
+    secondsRemaining > 0
+      ? `Please wait ${secondsRemainingText} and then restart.`
+      : "Please restart.";
+  return `You are not allowed to start a new Season 2 run so soon after ${action}. ${secondSentence}`;
 }
 
 function drawStartingRoomText() {
@@ -101,61 +103,4 @@ function drawStartingRoomText() {
     position.Y,
     KColorDefault,
   );
-}
-
-function getPosition(spriteName: keyof typeof sprites) {
-  const topLeftPositionGame = g.r.GetGridPosition(TOP_LEFT_GRID_INDEX);
-  const topLeftPosition = Isaac.WorldToRenderPosition(topLeftPositionGame);
-  const topRightPositionGame = g.r.GetGridPosition(TOP_RIGHT_GRID_INDEX);
-  const topRightPosition = Isaac.WorldToRenderPosition(topRightPositionGame);
-
-  switch (spriteName) {
-    case "characterTitle": {
-      return topLeftPosition.add(SPRITE_TITLE_OFFSET);
-    }
-
-    case "seededStartingTitle": {
-      return topRightPosition.add(SPRITE_TITLE_OFFSET);
-    }
-
-    case "seededItemCenter": {
-      return topRightPosition;
-    }
-
-    case "seededItemLeft": {
-      return topRightPosition.add(Vector(SPRITE_ITEM_OFFSET * -1, 0));
-    }
-
-    case "seededItemRight": {
-      return topRightPosition.add(Vector(SPRITE_ITEM_OFFSET, 0));
-    }
-
-    case "seededItemFarLeft": {
-      return topRightPosition.add(Vector(SPRITE_ITEM_OFFSET * -3, 0));
-    }
-
-    case "seededItemFarRight": {
-      return topRightPosition.add(Vector(SPRITE_ITEM_OFFSET * 3, 0));
-    }
-
-    default: {
-      return error(
-        `Starting room sprites named "${spriteName}" are unsupported.`,
-      );
-    }
-  }
-}
-
-function getSeason2ErrorMessage(action: string, secondsRemaining: int) {
-  if (secondsRemaining > SEASON_2_LOCK_SECONDS) {
-    return 'Please set your item vetos for Season 2 again in the "Change Char Order" custom challenge.';
-  }
-
-  const suffix = secondsRemaining > 1 ? "s" : "";
-  const secondsRemainingText = `${secondsRemaining} second${suffix}`;
-  const secondSentence =
-    secondsRemaining > 0
-      ? `Please wait ${secondsRemainingText} and then restart.`
-      : "Please restart.";
-  return `You are not allowed to start a new Season 2 run so soon after ${action}. ${secondSentence}`;
 }