+++ src/features/speedrun/season3/callbacks/postNewRoom.ts
@@ -1,17 +1,17 @@
 import { ChallengeCustom } from "../../../../enums/ChallengeCustom";
 import { getNumRoomsEntered } from "../../../utils/numRoomsEntered";
-import { resetSeason2StartingRoomSprites } from "../startingRoomSprites";
+import { resetSeason3StartingRoomSprites } from "../startingRoomSprites";
 
-export function season2PostNewRoom(): void {
+export function season3PostNewRoom(): void {
   const challenge = Isaac.GetChallenge();
 
-  if (challenge !== ChallengeCustom.SEASON_2) {
+  if (challenge !== ChallengeCustom.SEASON_3) {
     return;
   }
 
   const numRoomsEntered = getNumRoomsEntered();
 
   if (numRoomsEntered !== 1) {
-    resetSeason2StartingRoomSprites();
+    resetSeason3StartingRoomSprites();
   }
 }